+++ tests/Fixtures/User.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Mivento\FilamentSpreadsheetEditor\Tests\Fixtures;
+
+use Illuminate\Foundation\Auth\User as Authenticatable;
+
+class User extends Authenticatable
+{
+    protected $guarded = [];
+}